+++ tests/compat_device.c
@@ -21,7 +21,7 @@
  *  IN THE SOFTWARE.
  */
 
-#include "nng_compat.h"
+#include <nanomsg/nn.h>
 #include "compat_testutil.h"
 
 #define SOCKET_ADDRESS_A    "inproc://a"
