+++ tests/compat_shutdown.c
@@ -21,7 +21,7 @@
     IN THE SOFTWARE.
 */
 
-#include "nng_compat.h"
+#include <nanomsg/nn.h>
 #include "compat_testutil.h"
 
 int main (int argc, const char *argv[])
