+++ tests/compat_survey.c
@@ -21,7 +21,7 @@
    DEALINGS IN THE SOFTWARE.
 */
 
-#include "nng_compat.h"
+#include <nanomsg/nn.h>
 
 #include "compat_testutil.h"
